Work on #2196
Fixes #2196

All logic has moved to `AddWebFeedViewModel`.
